fix some comments, some files with mixed line endings, some warnings.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@757 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/State.cpp
@@ -185,7 +185,7 @@ void LoadStateCallback(u64 userdata, int cyclesLate)
 		}
 	} else {
 		fseek(f, 0, SEEK_END);
-		sz = ftell(f) - sizeof(int);
+		sz = (int)(ftell(f) - sizeof(int));
 		fseek(f, sizeof(int), SEEK_SET);
 
 		buffer = new u8[sz];
